+++ FrontEnd/src/Components/ControlPanel/Profesores/HeaderProfesores.jsx
@@ -1,4 +1,4 @@
-import React, { useState, useEffect, useMemo } from 'react';
+import React from 'react';
 import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import {
   Button,
