+++ FrontEnd/src/Pages/ControlPanel/PagesCP/profesores/Profesores.jsx
@@ -1,50 +1,38 @@
-import React, { useState, useEffect, useMemo } from 'react';
-import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
-import {
-  Button,
-  Modal,
-  TextField,
-  Box,
-  Typography,
-} from '@mui/material';
-import { grey } from '@mui/material/colors';
-import { DataGrid, gridClasses } from '@mui/x-data-grid';
-import Card from '@mui/material/Card';
-import CardContent from '@mui/material/CardContent';
-import Select from 'react-select';
-import { CSVLink } from 'react-csv';
-import Actions from './ActProfes';
+import React, { useState, useEffect } from 'react';
 import {
   createProfesor, deleteProfesor, getProfesors, updateProfesor,
 } from '../../../../api/profesors';
 import { getClasses } from '../../../../api/classes';
 import { getPeriodos } from '../../../../api/Periodos';
 import HeaderProfesores from '../../../../Components/ControlPanel/Profesores/HeaderProfesores';
+import ModalProfesor from '../../../../Components/ControlPanel/Profesores/ModalProfesor';
+import { profesorVacio } from '../../../../utils/constants';
+import BodyProfesores from '../../../../Components/ControlPanel/Profesores/BodyProfesores';
 
-export default function Profesores() {
-  let array = [];
-  let array2 = [];
-  let array3 = [];
-  const [data, setData] = useState([]);
-  const [guardaData, setGuardaData] = useState([]);
+function Profesores() {
+  const [profesorData, setProfesorData] = useState([]);
+  const [originalProfesorData, setOriginalProfesorData] = useState([]);
   const [dataPeriodo, setDataPeriodo] = useState([]);
   const [dataClase, setDataClase] = useState([]);
+  const [profesorSeleccionado, setProfesorSeleccionado] = useState(profesorVacio);
+  const [openModal, setOpenModal] = useState(false);
+  const [currentOperation, setCurrentOperation] = useState('');
 
   const getProfesores = async () => {
     await getProfesors().then((response) => response.json()).then((result) => {
-      setData(result);
+      setProfesorData(result);
+    });
+  };
+
+  const getProfesores2 = async () => {
+    await getProfesors().then((response) => response.json()).then((result) => {
+      setOriginalProfesorData(result);
     });
   };
 
   const getClases = async () => {
     await getClasses().then((response) => response.json()).then((result) => {
       setDataClase(result);
-    });
-  };
-
-  const getProfesores2 = async () => {
-    await getProfesors().then((response) => response.json()).then((result) => {
-      setGuardaData(result);
     });
   };
 
@@ -61,150 +49,68 @@ export default function Profesores() {
     getClases();
   }, []);
 
-  const [modalInsertar, setModalInsertar] = useState(false);
-  const [modalEditar, setModalEditar] = useState(false);
-  const [modalEliminar, setModalEliminar] = useState(false);
-
-  const [nombre, setNombre] = useState('');
-  const [apellidos, setApellidos] = useState('');
-  const [matricula, setMatricula] = useState('');
-  const [correo, setCorreo] = useState('');
-  const [fecha_de_nacimiento, setFecha_de_nacimiento] = useState('');
-  const [num_telefono, setNum_telefono] = useState('');
-  const [num_cursos_impartidos, setNum_cursos_impartidos] = useState('');
-  const [idUser, setIdUser] = useState('');
-  const [consolaSeleccionada, setConsolaSeleccionada] = useState({
-    _id: '',
-    nombre: '',
-    apellidos: '',
-    matricula: '',
-    correo: '',
-    fecha_de_nacimiento: '',
-    num_telefono: '',
-    num_cursos_impartidos: '',
-    idUser: '',
-  });
-
-  const abrirCerrarModalInsertar = () => {
-    setModalInsertar(!modalInsertar);
-  };
-
-  const abrirCerrarModalEditar = () => {
-    setModalEditar(!modalEditar);
-  };
-
-  const abrirCerrarModalEliminar = () => {
-    setModalEliminar(!modalEliminar);
-  };
-
-  const postCrea = async (e) => {
-    e.preventDefault();
-    try {
-      await createProfesor({
-        nombre,
-        apellidos,
-        matricula,
-        correo,
-        fecha_de_nacimiento,
-        num_telefono,
-        num_cursos_impartidos,
-        idUser,
-      });
-      abrirCerrarModalInsertar();
-      getProfesores();
-      setNombre('');
-      setApellidos('');
-      setMatricula('');
-      setCorreo('');
-      setFecha_de_nacimiento('');
-      setNum_telefono('');
-      setNum_cursos_impartidos('');
-      setIdUser('');
-    } catch (error) {
-      console.log(error);
+  const abrirCerrarModal = () => {
+    if (openModal) {
+      setProfesorSeleccionado(profesorVacio);
     }
+    setOpenModal(!openModal);
   };
 
   const handleChange = (e) => {
     const { name, value } = e.target;
-    setConsolaSeleccionada((prevState) => ({
+    setProfesorSeleccionado((prevState) => ({
       ...prevState,
       [name]: value,
     }));
   };
 
-  const seleccionarConsola = (consola, caso) => {
-    setConsolaSeleccionada(consola);
-    if (caso === 'Editar') {
-      abrirCerrarModalEditar();
-    } else {
-      abrirCerrarModalEliminar();
-    }
+  const seleccionarProfesor = (profesor, caso) => {
+    setProfesorSeleccionado(profesor);
+    setCurrentOperation(caso);
+    abrirCerrarModal();
   };
 
-  const postEditar = async (e) => {
+  const modalSubmit = async (e) => {
     e.preventDefault();
     try {
-      await updateProfesor({
-        _id: consolaSeleccionada._id,
-        nombre: consolaSeleccionada.nombre,
-        apellidos: consolaSeleccionada.apellidos,
-        matricula: consolaSeleccionada.matricula,
-        correo: consolaSeleccionada.correo,
-        fecha_de_nacimiento: consolaSeleccionada.fecha_de_nacimiento,
-        num_telefono: consolaSeleccionada.num_telefono,
-        num_cursos_impartidos: consolaSeleccionada.num_cursos_impartidos,
-        idUser: consolaSeleccionada.idUser,
-      }).then((response) => console.log(response));
-      abrirCerrarModalEditar();
-      getProfesores();
+      if (currentOperation === 'Eliminar') {
+        await deleteProfesor({
+          _id: profesorSeleccionado._id,
+        });
+      } else if (currentOperation === 'Crear') {
+        await createProfesor({
+          nombre: profesorSeleccionado.nombre,
+          apellidos: profesorSeleccionado.apellidos,
+          matricula: profesorSeleccionado.matricula,
+          correo: profesorSeleccionado.correo,
+          fecha_de_nacimiento: profesorSeleccionado.fecha_de_nacimiento,
+          num_telefono: profesorSeleccionado.num_telefono,
+          num_cursos_impartidos: profesorSeleccionado.num_cursos_impartidos,
+          idUser: profesorSeleccionado.idUser,
+        });
+      } else {
+        await updateProfesor({
+          _id: profesorSeleccionado._id,
+          nombre: profesorSeleccionado.nombre,
+          apellidos: profesorSeleccionado.apellidos,
+          matricula: profesorSeleccionado.matricula,
+          correo: profesorSeleccionado.correo,
+          fecha_de_nacimiento: profesorSeleccionado.fecha_de_nacimiento,
+          num_telefono: profesorSeleccionado.num_telefono,
+          num_cursos_impartidos: profesorSeleccionado.num_cursos_impartidos,
+          idUser: profesorSeleccionado.idUser,
+        });
+      }
     } catch (error) {
       console.log(error);
     }
+    abrirCerrarModal();
+    getProfesores();
   };
 
-  const postDelete = async (e) => {
-    try {
-      await deleteProfesor({
-        _id: consolaSeleccionada._id,
-      });
-      abrirCerrarModalEliminar();
-      getProfesores();
-    } catch (error) {
-      console.log(error);
-    }
-  };
-
-  const [pageSize, SetPageSize] = useState(5);
-
-  const columns = useMemo(
-    () => [
-      {
-        field: '_id', headerName: 'Id', width: 54, hide: true,
-      },
-      { field: 'nombre', headerName: 'Nombre', width: 120 },
-      { field: 'apellidos', headerName: 'Apellidos', width: 180 },
-      { field: 'matricula', headerName: 'Matricula', width: 140 },
-      { field: 'correo', headerName: 'Correo', width: 180 },
-      { field: 'num_telefono', headerName: 'Telefono', width: 120 },
-      { field: 'fecha_de_nacimiento', headerName: 'Nacimiento', width: 100 },
-      {
-        field: 'num_cursos_impartidos', headerName: 'Cursos Impartidos', width: 140,
-      },
-      {
-        field: 'idUser', headerName: 'Usuario', width: 250, hide: true,
-      },
-
-      {
-        field: 'actions',
-        headerName: 'Acciones',
-        type: 'actions',
-        width: 95,
-        renderCell: (params) => <Actions {...{ params, seleccionarConsola }} />,
-      },
-    ],
-    [data],
-  );
+  let array = [];
+  let array2 = [];
+  let array3 = [];
 
   const handleSelectChange = (event) => {
     array = [];
@@ -213,322 +119,43 @@ export default function Profesores() {
     array2.push(dataClase.filter((data) => data.clavePeriodo === event.label));
     for (let i = 0; i < array2.length; i++) {
       for (let j = 0; j < array2[i].length; j++) {
-        array.push(guardaData.filter((data) => data.matricula === array2[i][j].matriculaProfesor));
+        array.push(originalProfesorData.filter(
+          (data) => data.matricula === array2[i][j].matriculaProfesor,
+        ));
       }
     }
     for (let i = 0; i < array.length; i++) {
       array3.push(array[i][0]);
     }
     if (array3.length > 0) {
-      setData(array3);
+      setProfesorData(array3);
     } else {
       getProfesores();
     }
   };
 
-  const bodyInsertar = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 260,
-        height: 620,
-        backgroundColor: '#fefefd',
-        top: '48%',
-        left: '50%',
-        transform: 'translate(-48%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-      }}
-    >
-      <h3
-        style={{ paddingBottom: '15px', marginTop: '5px', fontFamily: 'arial' }}
-        align="center"
-      >
-        Agregar Profesor
-      </h3>
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Nombre"
-        onChange={(e) => setNombre(e.target.value)}
-        value={nombre}
-        autoFocus
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Apellidos"
-        onChange={(e) => setApellidos(e.target.value)}
-        value={apellidos}
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Matricula"
-        onChange={(e) => setMatricula(e.target.value)}
-        value={matricula}
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Correo"
-        onChange={(e) => setCorreo(e.target.value)}
-        value={correo}
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Fecha de nacimiento"
-        onChange={(e) => setFecha_de_nacimiento(e.target.value)}
-        value={fecha_de_nacimiento}
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Telefono"
-        onChange={(e) => setNum_telefono(e.target.value)}
-        value={num_telefono}
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Cursos impartidos"
-        onChange={(e) => setNum_cursos_impartidos(e.target.value)}
-        value={num_cursos_impartidos}
-      />
-      <br />
-      <br />
-      <div align="center">
-        <Button color="primary" onClick={postCrea}>
-          Insertar
-        </Button>
-        <Button onClick={abrirCerrarModalInsertar} color="error">
-          Cancelar
-        </Button>
-      </div>
-    </div>
-  );
-
-  const bodyEditar = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 260,
-        height: 620,
-        backgroundColor: '#fefefd',
-        top: '48%',
-        left: '50%',
-        transform: 'translate(-48%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-      }}
-    >
-      <h3
-        style={{ paddingBottom: '15px', marginTop: '5px', fontFamily: 'arial' }}
-        align="center"
-      >
-        Actualizar Profesor
-      </h3>
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Nombre"
-        onChange={handleChange}
-        value={consolaSeleccionada?.nombre}
-        name="nombre"
-        autoFocus
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Apellidos"
-        onChange={handleChange}
-        value={consolaSeleccionada?.apellidos}
-        name="apellidos"
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Matricula"
-        onChange={handleChange}
-        value={consolaSeleccionada?.matricula}
-        name="matricula"
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Correo"
-        onChange={handleChange}
-        value={consolaSeleccionada?.correo}
-        name="correo"
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Fecha de nacimiento"
-        onChange={handleChange}
-        value={consolaSeleccionada?.fecha_de_nacimiento}
-        name="fecha_de_nacimiento"
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Telefono"
-        onChange={handleChange}
-        value={consolaSeleccionada?.num_telefono}
-        name="num_telefono"
-      />
-      <br />
-      <TextField
-        style={{ paddingBottom: '15px', fontFamily: 'arial' }}
-        label="Cursos impartidos"
-        onChange={handleChange}
-        value={consolaSeleccionada?.num_cursos_impartidos}
-        name="num_cursos_impartidos"
-      />
-      <br />
-      <br />
-      <div align="center">
-        <Button color="primary" onClick={postEditar}>
-          Editar
-        </Button>
-        <Button onClick={abrirCerrarModalEditar} color="error">
-          Cancelar
-        </Button>
-      </div>
-    </div>
-  );
-
-  const bodyEliminar = (
-    <div
-      style={{
-        position: 'absolute',
-        width: 260,
-        height: 280,
-        backgroundColor: '#fefefd',
-        top: '50%',
-        left: '50%',
-        transform: 'translate(-50%, -50%)',
-        border: '4px solid  rgb(165, 165, 180)',
-        margin: 'auto',
-        borderRadius: '10px',
-        padding: '20px',
-      }}
-    >
-      <h3
-        style={{ paddingBottom: '15px', marginTop: '5px', fontFamily: 'arial' }}
-        align="center"
-      >
-        Eliminar una clase
-      </h3>
-      <Typography style={{ align: 'justify', fontFamily: 'arial' }}>
-        {`El profesor de ${consolaSeleccionada?.nombre} y
-        todo lo relacionado a el se va a eliminar por completo. No vas a poder
-        acceder a estos datos de nuevo.`}
-      </Typography>
-      <br />
-      <br />
-      <div align="center">
-        <Button color="error" onClick={postDelete}>
-          Confirmar
-        </Button>
-        <Button onClick={() => abrirCerrarModalEliminar()} color="primary">
-          Cancelar
-        </Button>
-      </div>
-    </div>
-  );
-
-  const [items, setItems] = useState([]);
   return (
     <div>
-      <HeaderProfesores 
-        data={data}
-        setOpenModal={abrirCerrarModalInsertar}
+      <HeaderProfesores
+        data={profesorData}
+        setOpenModal={() => { seleccionarProfesor(profesorVacio, 'Crear'); }}
         dataPeriodo={dataPeriodo}
         handleSelectChange={handleSelectChange}
       />
-      <Card
-        sx={{
-          width: 1120,
-          position: 'absolute',
-          textAlign: 'left',
-          marginLeft: '65px',
-          marginTop: '125px',
-          bgcolor: 'grey.200',
-          borderRadius: '8px',
-        }}
-      >
-        <CardContent>
-          <Typography
-            gutterBottom
-            variant="h5"
-            component="div"
-            sx={{ textAlign: 'left', fontFamily: 'arial', marginLeft: 1 }}
-          >
-            Filtro
-          </Typography>
-          <TextField
-            style={{
-              paddingBottom: '10px', fontFamily: 'arial', width: 1070, marginLeft: 7,
-            }}
-            label="Ingrese un nombre para buscar"
-            onChange={(e) => {
-              setItems([
-                {
-                  columnField: 'nombre',
-                  operatorValue: 'contains',
-                  value: e.target.value,
-                },
-              ]);
-            }}
-          />
-        </CardContent>
-      </Card>
-      <Box
-        sx={{
-          width: '1150px',
-          padding: '15px',
-          height: '450px',
-          position: 'absolute',
-          marginLeft: '50px',
-          marginTop: '270px',
-        }}
-      >
-        <DataGrid
-          columns={columns}
-          rows={data}
-          getRowId={(row) => row._id}
-          rowsPerPageOptions={[5, 10]}
-          pageSize={pageSize}
-          onPageSizeChange={(newPageSize) => SetPageSize(newPageSize)}
-          getRowSpacing={(params) => ({
-            top: params.isFirstVisible ? 0 : 5,
-            bottom: params.isLastVisible ? 0 : 5,
-          })}
-          sx={{
-            [`& .${gridClasses.row}`]: {
-              bgcolor: (theme) => (theme.palette.mode === 'light' ? grey[200] : grey[900]),
-              fontFamily: 'arial',
-            },
-          }}
-          disableSelectionOnClick
-          filterModel={{
-            items,
-          }}
-        />
-      </Box>
-      <Modal open={modalInsertar} onClose={abrirCerrarModalInsertar}>
-        {bodyInsertar}
-      </Modal>
-      <Modal open={modalEditar} onClose={abrirCerrarModalEditar}>
-        {bodyEditar}
-      </Modal>
-      <Modal open={modalEliminar} onClose={abrirCerrarModalEliminar}>
-        {bodyEliminar}
-      </Modal>
+      <BodyProfesores
+        data={profesorData}
+        seleccionarConsola={seleccionarProfesor}
+      />
+      <ModalProfesor
+        handleChange={handleChange}
+        consolaSeleccionada={profesorSeleccionado}
+        onSubmit={modalSubmit}
+        openModal={openModal}
+        abrirCerrarModal={abrirCerrarModal}
+        operation={currentOperation}
+      />
     </div>
   );
 }
+
+export default Profesores;
